*added* schema *removed* encoder, includer, validator, ... (those interfaces are now implemented by the schema)
Due to a lot of redudant code when writing encoders, includers and validators, I decided to follow a schema approach
similar to other serialization frameworks. However, the py-jsonapi schema should be (and is) designed to to create
jsonapi.org APIs.

Note: This is only an unstable, not working snapshot. Every components are there, we only need to wire everything up.